pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "matplotlib>=3.10.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "reportlab>=4.3.1",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
]
